Refactored code to use new entity WeatherInfo and use_case weather/WeatherInfoGetter
