+++ src/main/java/use_case/editTask/EditTaskInteractor.java
@@ -2,6 +2,7 @@ package use_case.editTask;
 
 import entity.Task;
 import entity.TaskInfo;
+import entity.WeatherInfo;
 import interface_adapter.addTask.Constants;
 import java.io.IOException;
 import java.time.LocalDate;
@@ -10,6 +11,8 @@ import java.util.List;
 import java.util.Map;
 import data_access.LocationService;
 import data_access.WeatherApiService;
+import org.jetbrains.annotations.NotNull;
+import use_case.weather.WeatherInfoGetter;
 
 /**
  * Interactor for the editTask use case
@@ -67,48 +70,9 @@ public class EditTaskInteractor implements EditTaskInputBoundary {
         }
 
         // Update weather info based on updated start time
-        String description = "";
-        String feels = "";
-        String iconName = "";
+        WeatherInfo weatherInfo = WeatherInfoGetter.getWeatherInfo(weatherApiService, info.getStartDateTime());
 
-        try {
-            LocalDateTime startDateTime = info.getStartDateTime();
-            LocalDate date = startDateTime.toLocalDate();
-            int hour = startDateTime.getHour();
-
-            List<Map<String,String>> hourly = weatherApiService.getHourlyWeather(LocationService.getUserCity(),
-                    date, hour, hour);
-
-            if (!hourly.isEmpty()) {
-                Map<String, String> hourlyMap = hourly.get(0);
-                description = hourlyMap.get("description");
-                feels = hourlyMap.get("feels");
-                Map<String, Object> daily = weatherApiService.getDailyWeather(LocationService.getUserCity(),
-                        date);
-                iconName = daily.get("iconName") != null ? daily.get("iconName").toString() : "";
-            }
-        } catch (IOException e) {
-            System.err.println("Weather Lookup Failed: " + e.getMessage());
-        }
-        String temp = feels;
-
-        // Create a new TaskInfo with updated weather data but preserve ID and deletion flag and status
-        TaskInfo updatedTaskInfo = new TaskInfo(
-                info.getId(),
-                info.getTaskName(),
-                info.getStartDateTime(),
-                info.getEndDateTime(),
-                info.getPriority(),
-                info.getTag(),
-                info.getReminder(),
-                info.getIsDeleted(),
-                description,
-                iconName,
-                temp
-        );
-        updatedTaskInfo.setTaskStatus(info.getTaskStatus());
-
-        Task taskToUpdate = new Task(updatedTaskInfo);
+        Task taskToUpdate = getTask(info, weatherInfo);
 
         // Update in data access
         dataAccess.updateTask(username, taskToUpdate);
@@ -118,5 +82,25 @@ public class EditTaskInteractor implements EditTaskInputBoundary {
                 false
         );
         presenter.prepareSuccessView(outputData);
+    }
+
+    @NotNull
+    private static Task getTask(TaskInfo info, WeatherInfo weatherInfo) {
+        TaskInfo updatedTaskInfo = new TaskInfo(
+                info.getId(),
+                info.getTaskName(),
+                info.getStartDateTime(),
+                info.getEndDateTime(),
+                info.getPriority(),
+                info.getTag(),
+                info.getReminder(),
+                info.getIsDeleted(),
+                weatherInfo.description(),
+                weatherInfo.iconName(),
+                weatherInfo.temperature()
+        );
+        updatedTaskInfo.setTaskStatus(info.getTaskStatus());
+
+        return new Task(updatedTaskInfo);
     }
 }
